pyproject.toml
[tool.poetry]
name = "coralsnake"
version = "0.0.3"
description = "MAPPING"
authors = ["Chang Ye <yech1990@gmail.com>"]
readme = "README.md"
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
dnaio = "^1.2.1"
rich-click = "^1.8.3"
pyfaidx = "^0.8.1.1"
pysam = "^0.22.1"

[tool.poetry.scripts]
coralsnake = 'coralsnake.cli:cli'

[build-system]
requires = ["poetry-core", "Cython", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
generate-setup-file = false
script = 'build_package.py'
